Support both unquoted path and path with options
EDITOR environment variable value may be:

- "'quoted path' -with -options"
- "unquoted path"

This previous patch supported only the first; this patch adds support to
for both options.

If an environment value is set, but the value cannot be resolved, the
user has to fix the configuration. There are two possible failures:

The environment variable value does not resolve to a file:

    Error: Invalid 'EDITOR' environment variable "foo bar": Cannot find editor.

Splitting the environment variable failed:

    Error: Invalid 'EDITOR' environment variable "foo bar' -with -options": No closing quotation.