Add -f to include files even if ignored
Users might have a .gitignore configuration where they ignore folders
like "Lib", "Build" "Misc" or other paths that might lead to blurb not
being able to add the files using just "git add". This change makes git
to "force" the addition of the files passed in, ensuring that even if
present in gitignore they will be added to the commit, rather than blurb
just failing in the operation.