Validate a branch that we parse when running cherry_picker --continue
When running cherry_picker --continue we count on being able to get
certain information from the branch's name which cherry_picker
constructed earlier.  Verify as best we can that the branch name is one
which cherry_picker could have constructed.

Relies on the changes here: https://github.com/python/core-workflow/pull/265
(which does the work of one of the validations)